+++ src/services/notification-queue-service/base-notification-replication-queue-service.ts
@@ -1,0 +1,8 @@
+import type { BaseNotificationTypeConfig } from '../../types/notification-type-config';
+
+export interface BaseNotificationReplicationQueueService<Config extends BaseNotificationTypeConfig> {
+  enqueueReplication(
+    notificationId: Config['NotificationIdType'],
+    backendIdentifier: string,
+  ): Promise<void>;
+}